backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2 ; python_version >= '3.7'
django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==9.3.0
pyjwt==2.6.0 ; python_version >= '3.7'
pytz==2022.6
sqlparse==0.4.3 ; python_version >= '3.5'
